Adapt three out of four remaining tests
